Migrating xrc to electrum based APIs

+++ FreeMarketOne.Search/XRCTransactionSummary.cs
@@ -4,8 +4,8 @@ namespace FreeMarketOne.Search
 {
     public class XRCTransactionSummary
     {
-        public decimal Total { get; set; }
-        public int Confirmations { get; set; }
+        public double Total { get; set; }
+        public long Confirmations { get; set; }
         public DateTimeOffset Date { get; set; }
     }
 }
